+++ .vscode/settings.json
@@ -12,5 +12,8 @@
   //style
   "css.validate": false,
   "less.validate": false,
-  "scss.validate": false
+  "scss.validate": false,
+  "cSpell.words": [
+    "Signin"
+  ]
 }
