+++ src/features/auth/components/SignIn/SignIn.tsx
@@ -1,12 +1,11 @@
-import { Controller, SubmitHandler } from 'react-hook-form';
 import { useRouter } from 'next/router';
+import { Controller, SubmitHandler } from 'react-hook-form';
 
-import TextField from '../reusable/TextField';
 import * as S from './SignIn.styles';
 import { SignInProps } from './SignIn.types';
-import Button from '../reusable/Button';
 
-import useSigninForm, { SigninFormType } from '@src/hooks/useSigninForm';
+import useSigninForm, { SigninFormType } from '@src/features/auth/hooks/useSigninForm';
+import { Button, TextField } from '@src/shared/components';
 
 const SignIn = ({}: SignInProps) => {
   const { control, formState, handleSubmit } = useSigninForm();
