+++ src/features/auth/components/SigninComplete/SigninComplete.tsx
@@ -2,13 +2,13 @@ import { useEffect, useRef } from 'react';
 import Lottie from 'lottie-web';
 import { useRouter } from 'next/router';
 
-import { SignupCompleteProps } from './SignupComplete.types';
-import * as S from './SignupComplete.styles';
-import Button from '../reusable/Button';
+import * as S from './SigninComplete.styles';
+import { SigninCompleteProps } from './SigninComplete.types';
 
-import signupLottie from '@src/assets/lotti/signup.json';
+import loginLottie from '@src/assets/lotti/login.json';
+import { Button } from '@src/shared/components';
 
-const SignupComplete = ({}: SignupCompleteProps) => {
+const SigninComplete = ({}: SigninCompleteProps) => {
   const lottieRef = useRef<HTMLDivElement>(null);
   const router = useRouter();
 
@@ -22,7 +22,7 @@ const SignupComplete = ({}: SignupCompleteProps) => {
       renderer: 'svg',
       loop: false,
       autoplay: true,
-      animationData: signupLottie,
+      animationData: loginLottie,
       path: '',
     });
   }, []);
@@ -35,7 +35,7 @@ const SignupComplete = ({}: SignupCompleteProps) => {
   return (
     <S.Wrapper>
       <S.Header>
-        <S.Title>회원가입을 완료하였습니다!</S.Title>
+        <S.Title>로그인을 완료하였습니다!</S.Title>
       </S.Header>
       <S.Content ref={lottieRef} />
       <S.ButtonContainer>
@@ -45,4 +45,4 @@ const SignupComplete = ({}: SignupCompleteProps) => {
   );
 };
 
-export default SignupComplete;
+export default SigninComplete;
