+++ src/features/auth/components/Signup/index.ts
@@ -1,2 +1,2 @@
-export { default as SignUp } from './Signup';
+export { default } from './Signup';
 export type { SignupProps as signupProps } from './Signup.types';
