+++ src/features/auth/components/index.ts
@@ -1,0 +1,4 @@
+export { default as SignIn } from './SignIn';
+export { default as SigninComplete } from './SigninComplete';
+export { default as Signup } from './Signup';
+export { default as SignupComplete } from './SignupComplete';
