+++ src/features/home/components/home/Home.types.ts
