+++ src/features/home/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Home } from './home';
