+++ src/pages/signin/index.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next';
 
-import SignIn from '@src/components/SignIn';
+import { SignIn } from '@src/features/auth/components';
 
 const SignInPage: NextPage = () => {
   return <SignIn />;
