+++ src/pages/signup/complete.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from 'next';
 
-import SignupComplete from '@src/components/SignupComplete';
+import { SignupComplete } from '@src/features/auth/components';
 
 const Complete: NextPage = () => {
   return <SignupComplete />;
